+++ client/browser/src/libs/code_intelligence/code_intelligence.test.tsx
@@ -7,7 +7,7 @@ jest.mock('react-dom', () => ({
 
 import { uniqueId } from 'lodash'
 import renderer from 'react-test-renderer'
-import { from, NEVER, of, Subject, Subscription } from 'rxjs'
+import { BehaviorSubject, from, NEVER, of, Subject, Subscription } from 'rxjs'
 import { filter, skip, switchMap, take } from 'rxjs/operators'
 import { Services } from '../../../../../shared/src/api/client/services'
 import { Range } from '../../../../../shared/src/api/extension/types/range'
@@ -15,6 +15,7 @@ import { integrationTestContext } from '../../../../../shared/src/api/integratio
 import { Controller } from '../../../../../shared/src/extensions/controller'
 import { PlatformContextProps } from '../../../../../shared/src/platform/context'
 import { isDefined } from '../../../../../shared/src/util/types'
+import { MutationRecordLike } from '../../shared/util/dom'
 import { createGlobalDebugMount, createOverlayMount, FileInfo, handleCodeHost } from './code_intelligence'
 import { toCodeViewResolver } from './code_views'
 
@@ -288,6 +289,116 @@ describe('code_intelligence', () => {
             await decorated()
             expect(line.querySelectorAll('.line-decoration-attachment').length).toBe(1)
             expect(line.querySelector('.line-decoration-attachment')!.textContent).toEqual('test decoration 2')
+        })
+
+        test('removes code views and models', async () => {
+            const { services } = await integrationTestContext(undefined, {
+                roots: [],
+                editors: [],
+            })
+            const codeView1 = createTestElement()
+            codeView1.className = 'code'
+            const codeView2 = createTestElement()
+            codeView2.className = 'code'
+            const fileInfo: FileInfo = {
+                repoName: 'foo',
+                filePath: '/bar.ts',
+                commitID: '1',
+            }
+            const mutations = new BehaviorSubject<MutationRecordLike[]>([
+                { addedNodes: [document.body], removedNodes: [] },
+            ])
+            subscriptions.add(
+                handleCodeHost({
+                    mutations,
+                    codeHost: {
+                        name: 'test',
+                        check: () => true,
+                        codeViewResolvers: [
+                            toCodeViewResolver('.code', {
+                                dom: {
+                                    getCodeElementFromTarget: jest.fn(),
+                                    getCodeElementFromLineNumber: jest.fn(),
+                                    getLineNumberFromCodeElement: jest.fn(),
+                                },
+                                resolveFileInfo: codeView => of(fileInfo),
+                            }),
+                        ],
+                        selectionsChanges: () => of([]),
+                    },
+                    extensionsController: createMockController(services),
+                    showGlobalDebug: true,
+                    ...createMockPlatformContext(),
+                })
+            )
+            let editors = await from(services.editor.editors)
+                .pipe(
+                    skip(3),
+                    take(1)
+                )
+                .toPromise()
+            expect(editors).toEqual([
+                {
+                    editorId: 'editor#0',
+                    isActive: true,
+                    model: {
+                        languageId: 'typescript',
+                        text: undefined,
+                        uri: 'git://foo?1#/bar.ts',
+                    },
+                    resource: 'git://foo?1#/bar.ts',
+                    selections: [],
+                    type: 'CodeEditor',
+                },
+                {
+                    editorId: 'editor#1',
+                    isActive: true,
+                    model: {
+                        languageId: 'typescript',
+                        text: undefined,
+                        uri: 'git://foo?1#/bar.ts',
+                    },
+                    resource: 'git://foo?1#/bar.ts',
+                    selections: [],
+                    type: 'CodeEditor',
+                },
+            ])
+            expect(services.model.hasModel('git://foo?1#/bar.ts')).toBe(true)
+            // Simulate codeView1 removal
+            mutations.next([{ addedNodes: [], removedNodes: [codeView1] }])
+            // One editor should have been removed, model should still exist
+            editors = await from(services.editor.editors)
+                .pipe(
+                    skip(1),
+                    take(1)
+                )
+                .toPromise()
+            expect(editors).toEqual([
+                {
+                    editorId: 'editor#1',
+                    isActive: true,
+                    model: {
+                        languageId: 'typescript',
+                        text: undefined,
+                        uri: 'git://foo?1#/bar.ts',
+                    },
+                    resource: 'git://foo?1#/bar.ts',
+                    selections: [],
+                    type: 'CodeEditor',
+                },
+            ])
+            expect(services.model.hasModel('git://foo?1#/bar.ts')).toBe(true)
+            // Simulate codeView2 removal
+            mutations.next([{ addedNodes: [], removedNodes: [codeView2] }])
+            // Second editor and model should have been removed
+            editors = await from(services.editor.editors)
+                .pipe(
+                    skip(1),
+                    take(1)
+                )
+                .toPromise()
+            expect(editors).toEqual([])
+            expect(services.model.hasModel('git://foo?1#/bar.ts')).toBe(false)
         })
     })
 })