+++ cmd/repo-updater/repos/migrations.go
@@ -4,8 +4,10 @@ import (
 	"context"
 	"fmt"
 	"net/url"
+	"strings"
 	"time"
 
+	"github.com/goware/urlx"
 	"github.com/pkg/errors"
 	"github.com/sourcegraph/sourcegraph/pkg/extsvc/github"
 	"github.com/sourcegraph/sourcegraph/pkg/jsonc"
@@ -27,27 +29,26 @@ func (m Migration) Run(ctx context.Context, s Store) error {
 // without recourse to the now deprecated enabled column of a repository.
 //
 // This is done by:
-//  1. Explicitly adding enabled repos that would have been deleted to an explicit include list.
-//  2. Explicitly adding disabled repos that would have been added to an explicit exclude list.
-//  3. Removing the deprecated initialRepositoryEnablement field.
+//  1. Explicitly adding disabled repos that would have been added to an explicit exclude list.
+//  2. Removing the deprecated initialRepositoryEnablement field.
 //
 // This migration must be rolled-out together with the UI changes that remove the admin's
 // ability to explicitly enabled / disable individual repos.
 func EnabledStateDeprecationMigration(sourcer Sourcer, clock func() time.Time, kinds ...string) Migration {
 	return migrate(func(ctx context.Context, s Store) error {
-		const prefix = "migrate.repos-enabled-state-deprecation"
+		const prefix = "migrate.repos-enabled-state-deprecation:"
 
 		es, err := s.ListExternalServices(ctx, StoreListExternalServicesArgs{
 			Kinds: kinds,
 		})
 
 		if err != nil {
-			return errors.Wrapf(err, "%s.list-external-services", prefix)
+			return errors.Wrapf(err, "%s list-external-services", prefix)
 		}
 
 		srcs, err := sourcer(es...)
 		if err != nil {
-			return errors.Wrapf(err, "%s.list-sources", prefix)
+			return errors.Wrapf(err, "%s list-sources", prefix)
 		}
 
 		var sourced Repos
@@ -58,7 +59,31 @@ func EnabledStateDeprecationMigration(sourcer Sourcer, clock func() time.Time, k
 		}
 
 		if err != nil {
-			return errors.Wrapf(err, "%s.sources.list-repos", prefix)
+			if strings.Contains(err.Error(), "abuse detection") || strings.Contains(err.Error(), "rate limit") {
+				// If this has occurred, we do not have enough rate limit to
+				// list all of the repositories and perform the migration. If
+				// we were to return the error right now then Kubernetes or
+				// Docker would restart this process and the migration would
+				// run again. Each time we run, we are consuming more rate
+				// limit and thus we may become deadlocked for multiple hours
+				// waiting for this migration to run. This is believed to have
+				// occurred already, see https://github.com/sourcegraph/sourcegraph/issues/3590
+				// and the linked issue's discussion for more information.
+				//
+				// Here we take a dumb approach to mitigate the changes of this
+				// happening: we wait 15m which is likely to replenish enough
+				// of our rate limit to allow the migration to go through (note
+				// we need 1 API request per 100 repositories).
+				//
+				// TODO(tsenart): String error comparison here and time.Sleep
+				// is a super ugly / hacky approach but works. Ideally the
+				// underlying source ListRepos method would slow down when
+				// hitting rate limiting instead for just this use case.
+				log15.Error("migrate.repos-enabled-state-deprecation: rate limiting detected, waiting 15m before retrying", "error", err)
+				time.Sleep(15 * time.Minute)
+				log15.Error("migrate.repos-enabled-state-deprecation: restarting..")
+			}
+			return errors.Wrapf(err, "%s sources.list-repos", prefix)
 		}
 
 		stored, err := s.ListRepos(ctx, StoreListReposArgs{
@@ -67,12 +92,11 @@ func EnabledStateDeprecationMigration(sourcer Sourcer, clock func() time.Time, k
 		})
 
 		if err != nil {
-			return errors.Wrapf(err, "%s.store.list-repos", prefix)
+			return errors.Wrapf(err, "%s store.list-repos", prefix)
 		}
 
 		type service struct {
 			svc     *ExternalService
-			include Repos
 			exclude Repos
 		}
 
@@ -129,44 +153,24 @@ func EnabledStateDeprecationMigration(sourcer Sourcer, clock func() time.Time, k
 			return err
 		}
 
-		err = group(
-			func(r *Repo) bool { return r.Enabled },
-			func(s *service) *Repos { return &s.include },
-			diff.Deleted,
-		)
-
-		if err != nil {
-			return err
-		}
-
 		now := clock()
 		for _, e := range svcs {
 			if err = removeInitalRepositoryEnablement(e.svc, now); err != nil {
-				return errors.Wrapf(err, "%s.remove-initial-repository-enablement", prefix)
+				return errors.Wrapf(err, "%s remove-initial-repository-enablement", prefix)
 			}
 
 			if len(e.exclude) > 0 {
 				if err = e.svc.Exclude(e.exclude...); err != nil {
-					return errors.Wrapf(err, "%s.exclude", prefix)
+					return errors.Wrapf(err, "%s exclude", prefix)
 				}
 				e.svc.UpdatedAt = now
 
-				log15.Info(prefix+".exclude", "service", e.svc.DisplayName, "repos", len(e.exclude))
+				log15.Info(prefix+" exclude", "service", e.svc.DisplayName, "repos", len(e.exclude))
 			}
-
-			if len(e.include) > 0 {
-				if err = e.svc.Include(e.include...); err != nil {
-					return errors.Wrapf(err, "%s.include", prefix)
-				}
-				e.svc.UpdatedAt = now
-
-				log15.Info(prefix+".include", "service", e.svc.DisplayName, "repos", len(e.include))
-			}
-
 		}
 
 		if err = s.UpsertExternalServices(ctx, upserts...); err != nil {
-			return errors.Wrapf(err, "%s.upsert-external-services", prefix)
+			return errors.Wrapf(err, "%s upsert-external-services", prefix)
 		}
 
 		var deleted Repos
@@ -179,7 +183,7 @@ func EnabledStateDeprecationMigration(sourcer Sourcer, clock func() time.Time, k
 		}
 
 		if err = s.UpsertRepos(ctx, deleted...); err != nil {
-			return errors.Wrapf(err, "%s.upsert-repos", prefix)
+			return errors.Wrapf(err, "%s upsert-repos", prefix)
 		}
 
 		return nil
@@ -205,21 +209,21 @@ func removeInitalRepositoryEnablement(svc *ExternalService, ts time.Time) error 
 // migration to its explicit default.
 func GithubSetDefaultRepositoryQueryMigration(clock func() time.Time) Migration {
 	return migrate(func(ctx context.Context, s Store) error {
-		const prefix = "migrate.github-set-default-repository-query"
+		const prefix = "migrate.github-set-default-repository-query:"
 
 		svcs, err := s.ListExternalServices(ctx, StoreListExternalServicesArgs{
 			Kinds: []string{"github"},
 		})
 
 		if err != nil {
-			return errors.Wrapf(err, "%s.list-external-services", prefix)
+			return errors.Wrapf(err, "%s list-external-services", prefix)
 		}
 
 		now := clock()
 		for _, svc := range svcs {
 			var c schema.GitHubConnection
 			if err := jsonc.Unmarshal(svc.Config, &c); err != nil {
-				return fmt.Errorf("%s: external service id=%d config unmarshaling error: %s", prefix, svc.ID, err)
+				return fmt.Errorf("%s external service id=%d config unmarshaling error: %s", prefix, svc.ID, err)
 			}
 
 			if len(c.RepositoryQuery) != 0 {
@@ -228,7 +232,7 @@ func GithubSetDefaultRepositoryQueryMigration(clock func() time.Time) Migration 
 
 			baseURL, err := url.Parse(c.Url)
 			if err != nil {
-				return errors.Wrapf(err, "%s.parse-url", prefix)
+				return errors.Wrapf(err, "%s parse-url", prefix)
 			}
 
 			_, githubDotCom := github.APIRoot(NormalizeBaseURL(baseURL))
@@ -240,7 +244,7 @@ func GithubSetDefaultRepositoryQueryMigration(clock func() time.Time) Migration 
 
 			edited, err := jsonc.Edit(svc.Config, c.RepositoryQuery, "repositoryQuery")
 			if err != nil {
-				return errors.Wrapf(err, "%s.edit-json", prefix)
+				return errors.Wrapf(err, "%s edit-json", prefix)
 			}
 
 			svc.Config = edited
@@ -248,7 +252,7 @@ func GithubSetDefaultRepositoryQueryMigration(clock func() time.Time) Migration 
 		}
 
 		if err = s.UpsertExternalServices(ctx, svcs...); err != nil {
-			return errors.Wrapf(err, "%s.upsert-external-services", prefix)
+			return errors.Wrapf(err, "%s upsert-external-services", prefix)
 		}
 
 		return nil
@@ -260,21 +264,21 @@ func GithubSetDefaultRepositoryQueryMigration(clock func() time.Time) Migration 
 // migration to its explicit default.
 func GitLabSetDefaultProjectQueryMigration(clock func() time.Time) Migration {
 	return migrate(func(ctx context.Context, s Store) error {
-		const prefix = "migrate.gitlab-set-default-project-query"
+		const prefix = "migrate.gitlab-set-default-project-query:"
 
 		svcs, err := s.ListExternalServices(ctx, StoreListExternalServicesArgs{
 			Kinds: []string{"gitlab"},
 		})
 
 		if err != nil {
-			return errors.Wrapf(err, "%s.list-external-services", prefix)
+			return errors.Wrapf(err, "%s list-external-services", prefix)
 		}
 
 		now := clock()
 		for _, svc := range svcs {
 			var c schema.GitLabConnection
 			if err := jsonc.Unmarshal(svc.Config, &c); err != nil {
-				return fmt.Errorf("%s: external service id=%d config unmarshaling error: %s", prefix, svc.ID, err)
+				return fmt.Errorf("%s  external service id=%d config unmarshaling error: %s", prefix, svc.ID, err)
 			}
 
 			if len(c.ProjectQuery) != 0 {
@@ -285,7 +289,7 @@ func GitLabSetDefaultProjectQueryMigration(clock func() time.Time) Migration {
 
 			edited, err := jsonc.Edit(svc.Config, c.ProjectQuery, "projectQuery")
 			if err != nil {
-				return errors.Wrapf(err, "%s.edit-json", prefix)
+				return errors.Wrapf(err, "%s edit-json", prefix)
 			}
 
 			svc.Config = edited
@@ -293,7 +297,7 @@ func GitLabSetDefaultProjectQueryMigration(clock func() time.Time) Migration {
 		}
 
 		if err = s.UpsertExternalServices(ctx, svcs...); err != nil {
-			return errors.Wrapf(err, "%s.upsert-external-services", prefix)
+			return errors.Wrapf(err, "%s upsert-external-services", prefix)
 		}
 
 		return nil
@@ -306,21 +310,21 @@ func GitLabSetDefaultProjectQueryMigration(clock func() time.Time) Migration {
 // behaviour of mirroring all repos accessible to the configured token.
 func BitbucketServerSetDefaultRepositoryQueryMigration(clock func() time.Time) Migration {
 	return migrate(func(ctx context.Context, s Store) error {
-		const prefix = "migrate.bitbucketserver-set-default-repository-query"
+		const prefix = "migrate.bitbucketserver-set-default-repository-query:"
 
 		svcs, err := s.ListExternalServices(ctx, StoreListExternalServicesArgs{
 			Kinds: []string{"bitbucketserver"},
 		})
 
 		if err != nil {
-			return errors.Wrapf(err, "%s.list-external-services", prefix)
+			return errors.Wrapf(err, "%s list-external-services", prefix)
 		}
 
 		now := clock()
 		for _, svc := range svcs {
 			var c schema.BitbucketServerConnection
 			if err := jsonc.Unmarshal(svc.Config, &c); err != nil {
-				return fmt.Errorf("%s: external service id=%d config unmarshaling error: %s", prefix, svc.ID, err)
+				return fmt.Errorf("%s  external service id=%d config unmarshaling error: %s", prefix, svc.ID, err)
 			}
 
 			if len(c.RepositoryQuery) != 0 {
@@ -334,7 +338,7 @@ func BitbucketServerSetDefaultRepositoryQueryMigration(clock func() time.Time) M
 
 			edited, err := jsonc.Edit(svc.Config, c.RepositoryQuery, "repositoryQuery")
 			if err != nil {
-				return errors.Wrapf(err, "%s.edit-json", prefix)
+				return errors.Wrapf(err, "%s edit-json", prefix)
 			}
 
 			svc.Config = edited
@@ -342,7 +346,61 @@ func BitbucketServerSetDefaultRepositoryQueryMigration(clock func() time.Time) M
 		}
 
 		if err = s.UpsertExternalServices(ctx, svcs...); err != nil {
-			return errors.Wrapf(err, "%s.upsert-external-services", prefix)
+			return errors.Wrapf(err, "%s upsert-external-services", prefix)
+		}
+
+		return nil
+	})
+}
+
+// BitbucketServerUsernameMigration returns a Migration that changes all
+// configurations of BitbucketServer external services to explicitly have the
+// `username` setting set to the user defined in the `url`, if any.
+// This will only happen if the `username` fields is empty or unset.
+func BitbucketServerUsernameMigration(clock func() time.Time) Migration {
+	return migrate(func(ctx context.Context, s Store) error {
+		const prefix = "migrate.bitbucketserver-username-migration:"
+
+		svcs, err := s.ListExternalServices(ctx, StoreListExternalServicesArgs{
+			Kinds: []string{"bitbucketserver"},
+		})
+
+		if err != nil {
+			return errors.Wrapf(err, "%s list-external-services", prefix)
+		}
+
+		now := clock()
+		for _, svc := range svcs {
+			var c schema.BitbucketServerConnection
+			if err := jsonc.Unmarshal(svc.Config, &c); err != nil {
+				return errors.Errorf("%s  external service id=%d config unmarshaling error: %s", prefix, svc.ID, err)
+			}
+
+			if c.Username != "" {
+				continue
+			}
+
+			u, err := urlx.Parse(c.Url)
+			if err != nil {
+				return errors.Wrapf(err, "%s parse-url", prefix)
+			}
+
+			username := u.User.Username()
+			if username == "" {
+				continue
+			}
+
+			edited, err := jsonc.Edit(svc.Config, username, "username")
+			if err != nil {
+				return errors.Wrapf(err, "%s edit-json", prefix)
+			}
+
+			svc.Config = edited
+			svc.UpdatedAt = now
+		}
+
+		if err = s.UpsertExternalServices(ctx, svcs...); err != nil {
+			return errors.Wrapf(err, "%s upsert-external-services", prefix)
 		}
 
 		return nil