+++ shared/src/api/client/services/editorService.test.ts
@@ -1,9 +1,11 @@
 import { Selection } from '@sourcegraph/extension-api-types'
+import { noop } from 'lodash'
 import { from, of, Subscribable } from 'rxjs'
 import { first, map } from 'rxjs/operators'
 import { TestScheduler } from 'rxjs/testing'
+import * as sinon from 'sinon'
 import { CodeEditor, createEditorService, EditorService, getActiveCodeEditorPosition } from './editorService'
-import { TextModel } from './modelService'
+import { ModelService, TextModel } from './modelService'
 
 export function createTestEditorService(
     editors: Subscribable<readonly CodeEditor[]> = of([])
@@ -19,6 +21,7 @@ describe('EditorService', () => {
             scheduler().run(({ cold, expectObservable }) => {
                 const editorService = createEditorService({
                     models: cold<TextModel[]>('a', { a: [{ uri: 'u', text: 't', languageId: 'l' }] }),
+                    removeModel: jest.fn(),
                 })
                 editorService.addEditor({ type: 'CodeEditor', resource: 'u', selections: [], isActive: true })
                 expectObservable(
@@ -38,7 +41,10 @@ describe('EditorService', () => {
     })
 
     test('addEditor', async () => {
-        const editorService = createEditorService({ models: of([{ uri: 'u', text: 't', languageId: 'l' }]) })
+        const editorService = createEditorService({
+            models: of([{ uri: 'u', text: 't', languageId: 'l' }]),
+            removeModel: jest.fn(),
+        })
         const editor = editorService.addEditor({ type: 'CodeEditor', resource: 'u', selections: [], isActive: true })
         expect(editor.editorId).toEqual('editor#0')
         expect(
@@ -59,7 +65,10 @@ describe('EditorService', () => {
 
     describe('setSelections', () => {
         test('ok', async () => {
-            const editorService = createEditorService({ models: of([{ uri: 'u', text: 't', languageId: 'l' }]) })
+            const editorService = createEditorService({
+                models: of([{ uri: 'u', text: 't', languageId: 'l' }]),
+                removeModel: jest.fn(),
+            })
             const editor = editorService.addEditor({
                 type: 'CodeEditor',
                 resource: 'u',
@@ -86,14 +95,17 @@ describe('EditorService', () => {
             ).toEqual([SELECTIONS])
         })
         test('not found', () => {
-            const editorService = createEditorService({ models: of([]) })
+            const editorService = createEditorService({ models: of([]), removeModel: jest.fn() })
             expect(() => editorService.setSelections({ editorId: 'x' }, [])).toThrowError('editor not found: x')
         })
     })
 
     describe('removeEditor', () => {
         test('ok', async () => {
-            const editorService = createEditorService({ models: of([{ uri: 'u', text: 't', languageId: 'l' }]) })
+            const editorService = createEditorService({
+                models: of([{ uri: 'u', text: 't', languageId: 'l' }]),
+                removeModel: noop,
+            })
             const editor = editorService.addEditor({
                 type: 'CodeEditor',
                 resource: 'u',
@@ -108,13 +120,46 @@ describe('EditorService', () => {
             ).toEqual([])
         })
         test('not found', () => {
-            const editorService = createEditorService({ models: of([]) })
+            const editorService = createEditorService({ models: of([]), removeModel: noop })
             expect(() => editorService.removeEditor({ editorId: 'x' })).toThrowError('editor not found: x')
+        })
+
+        it('calls removeModel() when removing the last editor that references the model', async () => {
+            const removeModel = sinon.spy<ModelService['removeModel']>(noop)
+            const editorService = createEditorService({
+                models: of([{ uri: 'u', text: 't', languageId: 'l' }]),
+                removeModel,
+            })
+            const editor1 = editorService.addEditor({
+                type: 'CodeEditor',
+                resource: 'u',
+                selections: [],
+                isActive: true,
+            })
+            const editor2 = editorService.addEditor({
+                type: 'CodeEditor',
+                resource: 'u',
+                selections: [],
+                isActive: true,
+            })
+            editorService.removeEditor(editor1)
+            sinon.assert.notCalled(removeModel)
+            editorService.removeEditor(editor2)
+            expect(
+                await from(editorService.editors)
+                    .pipe(first())
+                    .toPromise()
+            ).toEqual([])
+            sinon.assert.calledOnce(removeModel)
+            sinon.assert.calledWith(removeModel, 'u')
         })
     })
 
     test('removeAllEditors', async () => {
-        const editorService = createEditorService({ models: of([{ uri: 'u', text: 't', languageId: 'l' }]) })
+        const editorService = createEditorService({
+            models: of([{ uri: 'u', text: 't', languageId: 'l' }]),
+            removeModel: noop,
+        })
         editorService.addEditor({ type: 'CodeEditor', resource: 'u', selections: [], isActive: true })
         editorService.removeAllEditors()
         expect(