+++ CHANGELOG.md
@@ -9,11 +9,13 @@
 
 All notable changes to Sourcegraph are documented in this file.
 
-## Unreleased
+## 3.4.0 (unreleased)
 
 ### Added
 
 ## Changed
+
+- Indexed search is now enabled by default for new Docker deployments. (#3540)
 
 ### Removed
 
@@ -21,6 +23,24 @@ All notable changes to Sourcegraph are documented in this file.
 
 - Removes corrupted archives in the searcher cache and tries to populate the cache again instead of returning an error.
 - Fixed a bug where search scopes would not get merged, and only the lowest-level list of search scopes would appear.
+- Removed corrupted archives in the searcher cache and tries to populate the cache again instead of returning an error.
+
+## 3.3.3
+
+### Fixed
+
+- Fixed bug that prevented "Find references" action from being completed in the activation checklist.
+
+## 3.3.2
+
+### Fixed
+
+- Fixed an issue where the default `bitbucketserver.repositoryQuery` would not be created on migration from older Sourcegraph versions. https://github.com/sourcegraph/sourcegraph/issues/3591
+- Fixed an issue where Sourcegraph would add deleted repositories to the external service configuration. https://github.com/sourcegraph/sourcegraph/issues/3588
+- Fixed an issue where a repo-updater migration would hit code host rate limits. https://github.com/sourcegraph/sourcegraph/issues/3582
+- The required `bitbucketserver.username` field of a [Bitbucket Server external service configuration](https://docs.sourcegraph.com/admin/external_service/bitbucketserver#configuration), if unset or empty, is automatically migrated to match the user part of the `url` (if defined). https://github.com/sourcegraph/sourcegraph/issues/3592
+- Fixed a panic that would occur in indexed search / the frontend when a search error ocurred. https://github.com/sourcegraph/sourcegraph/issues/3579
+- Fixed an issue where the repo-updater service could become deadlocked while performing a migration. https://github.com/sourcegraph/sourcegraph/issues/3590
 
 ## 3.3.1